extensions/stock-api-fastapi/requirements.txt
fastapi==0.115.3
numpy==1.26.4
pandas==2.2.1
pydantic==2.10.4
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-exporter-otlp-proto-http==1.28.2
